feat(renderer): GUI V0.8
refactor: remove debug prints and clean up rendering logic.
fix: prevent rotation on mouse drag and reset cube state before shuffle.
feat: add option to choose solver and allow selection between random or user-defined spin list
